requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
